Add solution to exercise 4.21

+++ part4/bloglist/controllers/blogs.js
@@ -25,8 +25,24 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({
+      error: 'Invalid token.'
+    })
+  }
+  const blog = await Blog.findById(request.params.id)
+  const user = await User.findById(decodedToken.id)
+  if (blog.user.toString() === user._id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+    user.blogs = user.blogs.filter((blog) => blog.toString() !== request.params.id)
+    await user.save()
+    response.status(204).end()
+  } else {
+    response.status(401).json({
+      error: 'Unauthorized action.'
+    })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
